+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 Version 1.2-SNAPSHOT
 -------------
 
+CHANGED:
+
+- update of parent version to 5
+- added new prefixes for unique constraints and foreign keys in enum DatabasePrefix
+- update of ent-api dependency to version 1.1
+- replaced hibernate with javax.persistence-api dependency 
+
 Version 1.1
 -------------